+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Create `market_plot()` using ggplot2.
 * Order `market_price()` by contract factor.
+* Set market names as character strings.
 
 # predictr 0.0.1